Added a "process" that just spins forever

+++ kernel/src/userspace/syscall.c
@@ -1,5 +1,4 @@
 #include "syscall.h"
-#include "userspace.h"
 #include <cpuid.h>
 #include <cpu/msr.h>
 
@@ -13,7 +12,7 @@ bool syscall_available() {
 
 void syscall_initialize() {
     msr_set(MSR_STAR, (msr_get(MSR_STAR) & 0xFFFFFFFF) | ((uint64_t) 0x8 << 32) | (((uint64_t) (0x10 | 3) << 48)));
-    msr_set(MSR_LSTAR, syscall_entry);
+    msr_set(MSR_LSTAR, (uint64_t) syscall_entry);
     msr_set(MSR_SFMASK, msr_get(MSR_SFMASK) | 0xfffffffe);
     msr_set(MSR_EFER, msr_get(MSR_EFER) | 1);
 }
